requirements.txt
flask>=2.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
wordcloud>=1.9.0
nltk>=3.8.0
stop-words>=2018.7.23
emoji>=2.0.0
telethon>=1.24.0
python-dotenv>=1.0.0
joblib>=1.3.0
tensorflow-macos>=2.16.0
